refactor(BA-5613): move assign-by-name to permission_controller domain
Move action, service method, and processor from group domain to
permission_controller domain for consistency with the RBAC endpoint
and role-based entity model.

- Move action file to services/permission_contoller/actions/
- Move service method to PermissionControllerService
- Register processor in PermissionControllerProcessors with RBAC validator
- Update adapter to use _processors.permission_controller
- Remove from GroupService/GroupProcessors
